Refactor to Use Current API
- refactor to instead of using forecast to use current from darksky api
