+++ BikeCast/app/src/main/java/agbytech/com/bikecast/Listeners/OnForecastChangeListener.java
@@ -1,0 +1,11 @@
+package agbytech.com.bikecast.Listeners;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+/**
+ * Created by Anthony-Parkour on 11/23/16.
+ */
+public interface OnForecastChangeListener {
+    public void onForecastChanged(JSONObject current) throws JSONException;
+}
